Moved tests from GeneratorSpec related to havingLengthsBetween to HavingLengthsBetweenSpec.
